Modify SubmitButton component- it takes value form prosp now. Move it to UI directory.
