+++ src/components/UI/SubmitButton/SubmitButton.css
@@ -1,0 +1,19 @@
+.form__button--submit {
+  background-color: #000;
+  border: solid 3px #fade42;
+  color:#fade42;
+  font-weight: bold;
+  text-transform: uppercase;
+  cursor: pointer;
+  width: 150px;
+  transition: all 0.5s;
+  padding: 15px;
+}
+.form__button--submit:hover {
+  background-color: #fada27;
+  border-color: #000;
+  color: #000;
+}
+.form__button--submit:focus {
+  outline: none;
+}
